+++ @types/twitch-js/index.d.ts
@@ -161,6 +161,11 @@ declare module 'twitch-js' {
     on(event: Event.Raid, listener: (raid: Raid) => void): void
     on(event: Event.Cheer, listener: (channel: string, userstate: UserState, message: string) => void): void
     on(event: Event.EmoteSets, listener: (setsList: string, sets: EmoteSets) => void): void
+    on(event: Event.Notices, listener: (channel: string, noticeId: string, message: string) => void): void
+    on(
+      event: Event.MessageDeleted,
+      listener: (channel: string, messageId: string, username: string, message: string) => void
+    ): void
 
     say(channel: string, message: string): void
     whisper(username: string, message: string): void